Move around the get_neighbors mess
