+++ opencog/rule-engine/ChainerUtils.cc
@@ -113,57 +113,4 @@ bool are_similar(const Handle& h1, const Handle& h2, bool strict_type_match)
     return false;
 }
 
-HandleSeq get_neighbors(const Handle& h, Type desiredLinkType)
-{
-    HandleSeq answer;
-    for (const LinkPtr& link : h->getIncomingSet())
-    {
-        if (classserver().isA(link->getType(), UNORDERED_LINK)) continue;
-        if (link->getOutgoingAtom(0) != h) continue;
-        if (link->getType() != desiredLinkType) continue;
-
-        for (const Handle& handle : link->getOutgoingSet()) {
-           if (handle == h) continue;
-           answer.emplace_back(handle);
-        }
-    }
-    return answer;
-}
-
-/* Tail-recursive helper function. We mark it static, so that
- * gcc can optimize this, i.e. call it without buying the stack
- * frame. */
-static void get_distant_neighbors_rec(const Handle& h,
-                                      UnorderedHandleSet& res,
-                                      int dist)
-{
-    res.insert(h);
-
-    // Recursive calls
-    if (dist != 0) {
-        // 1. Fetch incomings
-        for (const LinkPtr& in_l : h->getIncomingSet()) {
-            Handle in_h = in_l->getHandle();
-            if (res.find(in_h) == res.cend()) // Do not re-explore
-                get_distant_neighbors_rec(in_h, res, dist - 1);
-        }
-        // 2. Fetch outgoings
-        LinkPtr link = LinkCast(h);
-        if (link) {
-            for (const Handle& out_h : link->getOutgoingSet()) {
-                if (res.find(out_h) == res.cend()) // Do not re-explore
-                    get_distant_neighbors_rec(out_h, res, dist - 1);
-            }
-        }
-    }
-}
-
-UnorderedHandleSet get_distant_neighbors(const Handle& h, int dist)
-{
-    UnorderedHandleSet results;
-    get_distant_neighbors_rec(h, results, dist);
-    results.erase(h);
-    return results;
-}
-
 } // namespace OpenCog
