+++ opencog/rule-engine/backwardchainer/Target.cc
@@ -23,7 +23,7 @@
 
 #include <opencog/util/random.h>
 
-#include "../ChainerUtils.h"
+#include <opencog/atomutils/Neighbors.h>
 #include "Target.h"
 
 using namespace opencog;
@@ -122,7 +122,7 @@ unsigned int Target::rule_count(const Rule& r) const
 	if (htarget == Handle::UNDEFINED)
 		return 0;
 
-	HandleSeq q = get_neighbors(htarget, SET_LINK);
+	HandleSeq q = get_target_neighbors(htarget, SET_LINK);
 
 	return std::count(q.begin(), q.end(), hname);
 }
